gestion-potager/gateway/src/test/javascript/spec/app/entities/microservice/cycle-de-vie/cycle-de-vie.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { CycleDeVieService } from 'app/entities/microservice/cycle-de-vie/cycle-de-vie.service';
import { ICycleDeVie, CycleDeVie } from 'app/shared/model/microservice/cycle-de-vie.model';

describe('Service Tests', () => {
  describe('CycleDeVie Service', () => {
    let injector: TestBed;
    let service: CycleDeVieService;
    let httpMock: HttpTestingController;
    let elemDefault: ICycleDeVie;
    let expectedResult: ICycleDeVie | ICycleDeVie[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CycleDeVieService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new CycleDeVie(0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CycleDeVie', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new CycleDeVie()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CycleDeVie', () => {
        const returnedFromService = Object.assign(
          {
            vitesseDeCroissance: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CycleDeVie', () => {
        const returnedFromService = Object.assign(
          {
            vitesseDeCroissance: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CycleDeVie', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
